requirements.txt
schedule
selenium
python-dotenv
slack-sdk
